blobstore,lvol: support cluster size above 2GiB
Any occurrence of cluster size in blobstore was change from uint32_t
to uint64_t, in order to enable cluster sizes above 2GiB.

+++ include/spdk/lvol.h
@@ -46,7 +46,7 @@ struct spdk_lvol_store;
 struct spdk_lvol;
 
 struct spdk_lvs_opts {
-	uint32_t cluster_sz;
+	uint64_t cluster_sz;
 };
 
 typedef void (*spdk_lvs_op_with_handle_complete)(void *cb_arg, struct spdk_lvol_store *lvol_store,
